Code refactor
Some improvements, few bugs killed,

+++ isNumber.c
@@ -1,14 +1,14 @@
 #define _CRT_SECURE_NO_WARNINGS
-#include "calculator.h"
+#include "cstring.h"
 #include <stdbool.h>
+#include <string.h>
 
-/* SPRAWDZ CZY CIAG JEST LICZBA */
+/* Check if the given string is a number */
 bool isNumber(char* expression) {
-	//check if the given string is a number
-	int expLength = strlen(expression);
+	size_t expLength = strlen(expression);
 	if (expLength < 1) { return false; }
 
-	for (int i = 0; i < expLength; i++) {
+	for (size_t i = 0; i < expLength; i++) {
 		//check all combinations of digits stored in chars
 		//we have integers, and floating point numbers with dot
 		//if none of these conditions are true return false
